+++ modules/apps/fragment/fragment-test/src/testIntegration/java/com/liferay/fragment/test/util/FragmentCompositionTestUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.fragment.util;
+package com.liferay.fragment.test.util;
 
 import com.liferay.fragment.model.FragmentCollection;
 import com.liferay.fragment.model.FragmentComposition;
